UPSTREAM: sbs-battery: make writes to ManufacturerAccess optional
According to the Smart Battery Data Specification, the use
of ManufacturerAcess (register 0x0) is implementation-defined.
It appears that some batteries use writes to this register
in order to implement certain functionality, but others may
simply NAK all writes to it. As a result, write failures to
ManufacturerAccess should not be used as an indicator of
battery presence, nor as a failure to enter sleep mode.

The failed write access was seen with SANYO AP13J3K.

Cc: Brian Norris <briannorris@chromium.org>
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Sebastian Reichel <sre@kernel.org>
(cherry picked from commit 17c6d3979e5bbff1de36a4e89015fa09ac495d27)

BUG=chromium:848112
TEST=kevin and scarlet battery

Change-Id: I243b6db3e9a20f0e81a5942d61a7be305569255e
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1081796
Reviewed-by: Guenter Roeck <groeck@chromium.org>