BACKPORT: drm/i915/guc: Prepare for nonblocking execbuf submission
Currently the presumption is that the request construction and its
submission to the GuC are all under the same holding of struct_mutex. We
wish to relax this to separate the request construction and the later
submission to the GuC. This requires us to reserve some space in the
GuC command queue for the future submission. For flexibility to handle
out-of-order request submission we do not preallocate the next slot in
the GuC command queue during request construction, just ensuring that
there is enough space later.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Link: http://patchwork.freedesktop.org/patch/msgid/20160909131201.16673-17-chris@chris-wilson.co.uk
(cherry picked from commit dadd481bfe559f09bc3c42e9d552149c1cb84e6a)
from drm-intel-next-2016-09-19
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

 Conflicts: [Tarun: Ignore "e02757d91 drm/i915/guc: add engine mask to
		 GuC client & pass to GuC". Not needed]
	drivers/gpu/drm/i915/intel_guc.h

Change-Id: I1cea925f60910bb67e644091d79e70a54f3c37c4