BACKPORT: ath10k: add provision for Rx descriptor abstraction
There are slight differences in Rx hw descriptor information
among different chips. So far driver does not use those new
information for any functionalities, but there is one important
information which is available from QCA99X0 onwards to indicate
the number of bytes that hw padded at the begining of the rx
payload and this information is needed to undecap the rx
packet. Add an abstraction for Rx desc to make use of the
new desc information available. The callback that this patch
defines to retrieve the padding bytes will be used in follow-up
patch.

Signed-off-by: Vasanthakumar Thiagarajan <vthiagar@qti.qualcomm.com>
[Rename operations to hw_ops for other purposes]
Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
(cherry-picked from commit ae02c8719aab19bf311b6ce2881feb844456297e)

Conflicts:
	* no QCA9887, QCA9984, QCA9888 in ar10k
	* context, around missing .sw_decrypt_mcast_mgmt
	* struct ath10k_hw_params is in core.h, not hw.h

BUG=b:70048735
TEST=kernel still builds; QCA6174A Wifi still works

Change-Id: I19be60c0eb25a0ef77374282c44b9450f48416a4
Signed-off-by: Bartosz Chronowski <ext.bartosz.chronowski@tieto.com>
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/861624
Reviewed-by: Grant Grundler <grundler@chromium.org>