UPSTREAM: Bluetooth: btrtl: Add RTL8822BE Bluetooth device
The RTL8822BE is a new Realtek wifi and BT device. Support for the BT
part is hereby added.

As this device is similar to most of the other Realtek BT devices, the
changes are minimal. The main difference is that the 8822BE needs a
configuration file for enabling and disabling features. Thus code is
added to select and load this configuration file. Although not needed
at the moment, hooks are added for the other devices that might need
such configuration files.

One additional change is to the routine that tests that the project
ID contained in the firmware matches the hardware. As the project IDs
are not sequential, continuing to use the position in the array as the
expected value of the ID would require adding extra unused entries in
the table, and any subsequant rearrangment of the array would break the
code. To fix these problems, the array elements now contain both the
hardware ID and the expected value for the project ID.

BUG=chromium:704196
TEST=Deploy v5.44 BlueZ system daemon with v4.9 BlueZ kernel on Kevin and
perform automated tests (suite:bluetooth) and manual tests against mouses,
keyboards and headphones.
CQ-DEPEND=CL:474383

Change-Id: I29ba51eb490391bf534954c17e571eac80ed239a
Signed-off-by:  <alex_lu@realsil.com.cn>
Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Miao-chen Chou <mcchou@chromium.org>
(cherry picked from commit 1110a2dbe69831abdcf119c3a9a4c4ef2d0905f8)
Reviewed-on: https://chromium-review.googlesource.com/474342
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>