CHROMIUM: arm64: dts: rockchip: Flip assigned clock orders in VOPs
The VOPs specify some assigned clock rates for their clocks in their
device tree node.  Specifically:
* ACLK / HCLK: Nobody sets these rates but they need to be sane.
* DCLK: This is the pixel clock and will eventually be set, but needs to
  be at a sane rate during init time.

It appears that when we set these clocks up the order is quite
important, partially due to some of the hacks we have in
http://crosreview.com/362820.  Specifically:

1. At boot (if firmware doesn't touch) then ACLK, HCLK, and DCLK for
   VOP0 are all parented on VPLL.
2. Due to our hacks to the clock framework, when we change the DCLK we
   allow it to change VPLL to suit its needs.
3. If we change DCLK first then the ACLK and HCLK will temporarily have
   a screwy clock rate.

Let's avoid these problems by specifying the ACLK / HCLK first (which
should cause them to reparent to another PLL) and _then_ change
DPLL (which should change VPLL).

It's still a little unclear why glitching the clock rate of ACLK / HCLK
causes so many problems, but this definitely seems to avoid them.

BUG=chrome-os-partner:57351
TEST=Boot w/ gbb flags of 0x0 and see no hang.

Change-Id: I102294b88ba8e3c80e27b0fd7b40ba51c6da5886
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/384396
Reviewed-by: Guenter Roeck <groeck@chromium.org>