UPSTREAM: dma-buf/sync_file: free fences array in num_fences is 1
When merging sync_files there is a case when we can end up with only one
fence in the merged sync_file: when all fences belong to the same
timeline.

So for this case a fence_array is not created instead we just assigned the
fence to sync_file->fence. Then we do not use the fences array anymore nor
does free it.

This patch frees the array.

Reported-by:  Chris Wilson <chris@chris-wilson.co.uk>
BUG=chromium:911825
TEST=boot caroline, chell, run login_LoginSuccess, graphics_SanAngeles, cheets_StartAndroid

Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.co.uk>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/1474442419-6040-1-git-send-email-gustavo@padovan.org
(cherry picked from commit c654dd07522e88bacc546c9ccd81b3ee72838ccf)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: If82b9f3d91868a56208cc103da6cd0455d3e05b0