CHROMIUM: drm/img-rogue: Fixed logic in PVRSRVDevicePreClockSpeedChange
Added handling of the return value of PVRSRVDevicePreClockSpeedChange.

This fixes a case where the Host tries to put the GPU in forced idle state
(as part of PVRSRVDevicePreClockSpeedChange) but that request is denied,
then later on it tries to cancel the forced idle state
(as part of PVRSRVDevicePostClockSpeedChange).

A "cancel forced idle" request sent while the GPU is not forced idle is an
error (and causes an assert in the Firmware).

BUG=chrome-os-partner:57261
TEST=

Change-Id: I6d84582dd0d1cd437001f6502d2d4a2e3e91ce05
Signed-off-by: Jones Chen <jones.chen@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/388265
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>