UPSTREAM: drm/i915/gen9: implement missing case for SKL watermarks calculation
This should affect linear and X tiled planes on really small htotal
cases. It doesn't seem to be a very feasible case, but let's implement
it since it's on the specification and it's better to have it and
never need than not have it and realize we needed it.

Reviewed-by: Lyude <cpaul@redhat.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1474578035-424-9-git-send-email-paulo.r.zanoni@intel.com
(cherry picked from commit f1db3eafe5a2ad39caa8315f9b5403759e0c5520)
from drm-intel-next-2016-10-10
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

Change-Id: I2330d6dfe0e88778c76ba7dfdbeb7b5e81f57bde
Reviewed-on: https://chromium-review.googlesource.com/460762
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>