CHROMIUM: HACK: TO-REVERT: Lower clocks; disable features, bump voltages
This is a squash of a bunch of test patches that we may land to see if
it makes things more stable.  The expectation is that this will be
rather quickly reverted and if we want some of these fixes later we'll
re-add them.

Total changes:
* Chop off top Big/Little frequency (1.8 / 1.4 max)
* Increase Big/Little voltage by 100 mV
* CCI back down to 600 MHz (from 800 MHz)
* No CPU Idle
* Fix GPU glitch

The above is expected universally to:
* Increase power usage
* Decrease performance
* Increase stability

BUG=chrome-os-partner:57791
TEST=Confirm board gets updated OK

Change-Id: I1e05a6f31edd438c8ee06dc4fc0796d79fdbd55d
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/391241