CHROMIUM: iwl7000: pcie: introduce split point to a000 devices
a000 devices are going to have a lot of flows simplified
and changed: init flow, RX, TX, and more.
This, combined with the fact that code is already very
complicated due to backward compatibility - introduce
a split that will enable to introduce simplified version
of functions.
Shared ops are moved to a macro, while functions that will
be updated in the next patches are defined twice for now.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: b654fa203acf985edb1cf0cc5cb1fb72a844d058