CHROMIUM: dma-buf/fence-array: Signal the base fence from a worker
There is a problem with signaling the array fence in the same context
as the callback received from the last remaining fence in the array,
that might cause a deadlock with fence_add_callback() called on that
fence in parallel.

When the array fence callback is called, it calls fence_signal() on the
array fence, which in turn attempts to acquire the array fence lock, while
the thread is already holding the lock of the aggregated fence. At the
same time, if one calls fence_add_callback() on the same aggregated fence,
it will attempt to lock the aggregated fence and then call
fence_add_callback() on the array fence (through the array fence
.enable_signaling() callback), which requires the array fence lock. This
leads to a typical ABBA deadlock scenario.

Since all other fence operations should be happening from the top of the
aggregation (as opposed to signal, which originates at the bottom),
let's make the array fence signal happen from an asynchronous worker,
which doesn't hold the aggregated fence lock.

BUG=chrome-os-partner:58054
TEST=Enable KASAN and run webgl fishtank and Android In Car Racing

Change-Id: Icdc34da5ba65f925d4d1af1e5051cc8c80f270e6
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/394529
(cherry picked from commit 12b978c0d4291888e1d8cdc8c6d461c48c600a21)
Reviewed-on: https://chromium-review.googlesource.com/397030