UPSTREAM: HID: sony: Adjust value range for motion sensors
The motion sensor values are 16-bit, so make the value range match.
It is hard to reach the upper values, but they can be reached. At
least the current accelerometer value of 8192 is very easy to pass.

It is still not nice that the motion sensors live in no man's land
in between ABS_MISC and ABS_MT_SLOT, but that's something for another
time, which the proposed ABS_ACCEL_*/ABS_GYRO_* were meant for.

Signed-off-by: Roderick Colenbrander <roderick.colenbrander@sony.com>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit bdae9e0e95364123fb7d372872bd7efd1760867c)

BUG=b:139939892, chromium:972297
TEST=DualShock v4 via Bluetooth works on https://html5gamepad.com/

Change-Id: I91196b81ff3ea95840c6ce2d5cea1213d26a68cd
Signed-off-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1862779
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>