CHROMIUM: arm64: dts: rockchip: Set 99% settle times for rk3399-gru
The DVS rails were originally all set to a 200 us settle time because:
* They were calculated based on 90% rise.
* They were all given different amounts of slack to make the numbers
  look pleasing to the eye (200 is really pretty, isn't it?)

Let's change this to be based on 99% rise.  It's also silly to
arbitrarily give each value a different amount of slack. Now we'll
actually include the raw numbers here based on calculations.  It's
presumed that the fact that we're waiting for 99% rise is enough to give
a little bit of slack anyway.  These numbers are also the worse case
rise time.  If we really need extra slack, we should add a number that's
a percentage of the real value, not something to just make a nice round
number.

BUG=chrome-os-partner:58432
TEST=Stability test

Change-Id: Ie59a9d5f9fc3283cfb7c583054479798ccc32ae8
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/396961