CHROMIUM: config: arm64: panic on hardlockup detect
We've had this set to panic-on-hardlockup for a long time for
rk3288/arm32. Let's do that for rk3399/arm64 too.

This helps for debugging and collecting crash information, since
otherwise a hardlockup is not guaranteed to trigger the panic handlers
(which we recently instrumented for rk3399 to use ARM debug
infrastructure to dump info about recently-retired instruction
addresses).

BUG=chromium:654931, chrome-os-partner:58229
TEST=`echo HARDLOCKUP > /sys/kernel/debug/provoke-crash/DIRECT`; see
     panic, instead of indefinite hang

Change-Id: I45f407539d59ff6404e47ec774078874b4c63153
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/396959