UPSTREAM: drm/i915: Update debugfs describe_obj() to show fault-mappable
The current meaning of whether an object has a GGTT vma is very
ill-defined (and note we don't check for any partials either), it just
means that at some point it was in the GGTT but it may not be now. The
information we really care about here is whether it is taking up
precious mappable aperture space. This is the obj->fault_mappable flag.
We have a redundant long form reprinting of this information, so remove
that in favour of the compact flag.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20161012114827.17031-2-chris@chris-wilson.co.uk
(cherry picked from commit 8baa1f04b9ed8afe8d2d9f3884b335b1bab984d0)
from drm-intel-next-2016-10-24
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=No regressions in KBL and APL based devices

Change-Id: I3b99f2b01c8b8ef2a467856b4dc86e9817962cb3
Reviewed-on: https://chromium-review.googlesource.com/460791
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>